#25 setting: data 모듈 model 위치 변경
