+++ data/src/main/java/com/exhibitiondot/data/datasource/comment/CommentDataSource.kt
@@ -1,8 +1,8 @@
 package com.exhibitiondot.data.datasource.comment
 
 import androidx.paging.PagingData
-import com.exhibitiondot.data.model.dto.CommentDto
-import com.exhibitiondot.data.model.request.AddCommentRequest
+import com.exhibitiondot.data.network.model.dto.CommentDto
+import com.exhibitiondot.data.network.model.request.AddCommentRequest
 import com.exhibitiondot.data.network.NetworkState
 import kotlinx.coroutines.flow.Flow
