+++ data/src/main/java/com/exhibitiondot/data/datasource/comment/CommentPagingSource.kt
@@ -4,7 +4,7 @@ import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import com.exhibitiondot.data.network.api.CommentApi
 import com.exhibitiondot.data.network.api.ApiConst
-import com.exhibitiondot.data.model.dto.CommentDto
+import com.exhibitiondot.data.network.model.dto.CommentDto
 import com.exhibitiondot.data.network.NetworkState
 
 class CommentPagingSource(
