+++ data/src/main/java/com/exhibitiondot/data/datasource/event/EventDataSource.kt
@@ -1,8 +1,8 @@
 package com.exhibitiondot.data.datasource.event
 
 import androidx.paging.PagingData
-import com.exhibitiondot.data.model.dto.EventDetailDto
-import com.exhibitiondot.data.model.dto.EventDto
+import com.exhibitiondot.data.network.model.dto.EventDetailDto
+import com.exhibitiondot.data.network.model.dto.EventDto
 import com.exhibitiondot.data.network.NetworkState
 import com.exhibitiondot.domain.model.EventParams
 import kotlinx.coroutines.flow.Flow
