+++ data/src/main/java/com/exhibitiondot/data/datasource/user/UserRemoteDataSource.kt
@@ -1,11 +1,11 @@
 package com.exhibitiondot.data.datasource.user
 
 import com.exhibitiondot.data.network.api.UserApi
-import com.exhibitiondot.data.model.dto.UserDto
-import com.exhibitiondot.data.model.request.ChangeUserInfoRequest
-import com.exhibitiondot.data.model.request.SignInRequest
-import com.exhibitiondot.data.model.request.SignUpRequest
-import com.exhibitiondot.data.model.response.SignInResponse
+import com.exhibitiondot.data.network.model.dto.UserDto
+import com.exhibitiondot.data.network.model.request.ChangeUserInfoRequest
+import com.exhibitiondot.data.network.model.request.SignInRequest
+import com.exhibitiondot.data.network.model.request.SignUpRequest
+import com.exhibitiondot.data.network.model.response.SignInResponse
 import com.exhibitiondot.data.network.NetworkState
 import javax.inject.Inject
