+++ data/src/main/java/com/exhibitiondot/data/network/api/UserApi.kt
@@ -1,10 +1,10 @@
 package com.exhibitiondot.data.network.api
 
-import com.exhibitiondot.data.model.dto.UserDto
-import com.exhibitiondot.data.model.request.ChangeUserInfoRequest
-import com.exhibitiondot.data.model.request.SignInRequest
-import com.exhibitiondot.data.model.request.SignUpRequest
-import com.exhibitiondot.data.model.response.SignInResponse
+import com.exhibitiondot.data.network.model.dto.UserDto
+import com.exhibitiondot.data.network.model.request.ChangeUserInfoRequest
+import com.exhibitiondot.data.network.model.request.SignInRequest
+import com.exhibitiondot.data.network.model.request.SignUpRequest
+import com.exhibitiondot.data.network.model.response.SignInResponse
 import com.exhibitiondot.data.network.NetworkState
 import retrofit2.http.Body
 import retrofit2.http.GET
