+++ data/src/main/java/com/exhibitiondot/data/network/model/dto/EventDetailDto.kt
@@ -1,4 +1,4 @@
-package com.exhibitiondot.data.model.dto
+package com.exhibitiondot.data.network.model.dto
 
 import kotlinx.serialization.Serializable
