+++ data/src/main/java/com/exhibitiondot/data/network/model/request/ChangeUserInfoRequest.kt
@@ -1,12 +1,9 @@
-package com.exhibitiondot.data.model.request
+package com.exhibitiondot.data.network.model.request
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SignUpRequest(
-    val email: String,
-    val name: String,
-    val phone: String,
+data class ChangeUserInfoRequest(
     val nickname: String,
     val region: String,
     val categoryList: List<String>,
