+++ data/src/main/java/com/exhibitiondot/data/network/model/request/SignUpRequest.kt
@@ -1,9 +1,12 @@
-package com.exhibitiondot.data.model.request
+package com.exhibitiondot.data.network.model.request
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ChangeUserInfoRequest(
+data class SignUpRequest(
+    val email: String,
+    val name: String,
+    val phone: String,
     val nickname: String,
     val region: String,
     val categoryList: List<String>,
