+++ data/src/main/java/com/exhibitiondot/data/network/model/response/CommentListResponse.kt
@@ -1,6 +1,6 @@
-package com.exhibitiondot.data.model.response
+package com.exhibitiondot.data.network.model.response
 
-import com.exhibitiondot.data.model.dto.CommentDto
+import com.exhibitiondot.data.network.model.dto.CommentDto
 import kotlinx.serialization.Serializable
 
 @Serializable
