+++ data/src/main/java/com/exhibitiondot/data/repository/CommentRepositoryImpl.kt
@@ -4,8 +4,8 @@ import androidx.paging.PagingData
 import androidx.paging.map
 import com.exhibitiondot.data.datasource.comment.CommentDataSource
 import com.exhibitiondot.data.mapper.toDomain
-import com.exhibitiondot.data.model.dto.CommentDto
-import com.exhibitiondot.data.model.request.AddCommentRequest
+import com.exhibitiondot.data.network.model.dto.CommentDto
+import com.exhibitiondot.data.network.model.request.AddCommentRequest
 import com.exhibitiondot.data.network.NetworkState
 import com.exhibitiondot.domain.model.Comment
 import com.exhibitiondot.domain.repository.CommentRepository
