+++ data/src/main/java/com/exhibitiondot/data/repository/EventRepositoryImpl.kt
@@ -4,7 +4,7 @@ import androidx.paging.PagingData
 import androidx.paging.map
 import com.exhibitiondot.data.datasource.event.EventDataSource
 import com.exhibitiondot.data.mapper.toDomain
-import com.exhibitiondot.data.model.dto.EventDto
+import com.exhibitiondot.data.network.model.dto.EventDto
 import com.exhibitiondot.data.network.NetworkState
 import com.exhibitiondot.domain.model.Event
 import com.exhibitiondot.domain.model.EventDetail
